requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.1.1
fonttools==4.37.4
frozenlist==1.3.1
google-auth==2.12.0
google-auth-oauthlib==0.4.6
goprocam==4.2.0
grpcio==1.49.1
idna==3.4
importlib-metadata==5.0.0
ipython==8.5.0
jedi==0.18.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pyTelegramBotAPI==4.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
stack-data==0.5.1
subpub==1.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.9.0